Revert "ANDROID: Build script"
This reverts commit fafdaeec7b2b67a06c7523cdb5c6038063db04bb.

+++ jvm-packages/xgboost4j/src/native/xgboost4j.cpp
@@ -23,7 +23,7 @@
 // helper functions
 // set handle
 void setHandle(JNIEnv *jenv, jlongArray jhandle, void* handle) {
-  long out = (long) handle;
+  long long out = (long long) handle;
   jenv->SetLongArrayRegion(jhandle, 0, 1, &out);
 }
